split root view web/mobile
