+++ packages/@coorpacademy-app-review/package.json
@@ -7,7 +7,7 @@
     "node": ">=16.15.0"
   },
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "test:unit": "echo @todo setup unit tests"
   },
   "author": "CoorpAcademy",
   "files": [
